DAS-2067 - Migrate datatree/extensions.py.
